This implements the go to symbol feature.
It was already possible to jump to 'global' symbols, but this code should also allow to jumpt to all
locally defined definitions as well. This also means that those locally defined variables can be
renamed. In other words, for example function arguments should be renameable through the normal shortcut
and should leave the code compiling afterwards.